Replace DPI resolution searching with current Hi-DPI values
The existing function here tries to calculate the true native DPI of the monitor. It does this by searching all display modes, looking for a native mode. It then uses the "true" width and height values, rather than the "scaled" values to calculate DPI.
This isn't what we want for the IDE - we want the "scaled" resolution of whatever display mode the user has set.